refactor cli command & ui code with reuseable util functions
